Tons of fixes with workspace compositors

+++ external/terra/OgreHlmsPbsTerraShadows.cpp
@@ -39,94 +39,94 @@ THE SOFTWARE.
 
 namespace Ogre
 {
-    const IdString PbsTerraProperty::TerraEnabled   = IdString( "terra_enabled" );
+    const IdString PbsTerraProperty::TerraEnabled = IdString("terra_enabled");
 
     HlmsPbsTerraShadows::HlmsPbsTerraShadows() :
-          mTerra( 0 )
-        , mTerraSamplerblock( 0 )
+        mTerra(0)
+        , mTerraSamplerblock(0)
 #if OGRE_DEBUG_MODE
-        , mSceneManager( 0 )
+        , mSceneManager(0)
 #endif
     {
     }
     //-----------------------------------------------------------------------------------
     HlmsPbsTerraShadows::~HlmsPbsTerraShadows()
     {
-        if( mTerraSamplerblock )
+        if (mTerraSamplerblock)
         {
-            HlmsManager *hlmsManager = Root::getSingleton().getHlmsManager();
-            hlmsManager->destroySamplerblock( mTerraSamplerblock );
+            HlmsManager* hlmsManager = Root::getSingleton().getHlmsManager();
+            hlmsManager->destroySamplerblock(mTerraSamplerblock);
             mTerraSamplerblock = 0;
         }
     }
     //-----------------------------------------------------------------------------------
-    void HlmsPbsTerraShadows::setTerra( Terra *terra )
+    void HlmsPbsTerraShadows::setTerra(Terra* terra)
     {
         mTerra = terra;
-        if( !mTerraSamplerblock )
+        if (!mTerraSamplerblock)
         {
-            HlmsManager *hlmsManager = Root::getSingleton().getHlmsManager();
-            mTerraSamplerblock = hlmsManager->getSamplerblock( HlmsSamplerblock() );
+            HlmsManager* hlmsManager = Root::getSingleton().getHlmsManager();
+            mTerraSamplerblock = hlmsManager->getSamplerblock(HlmsSamplerblock());
         }
     }
     //-----------------------------------------------------------------------------------
-    uint16 HlmsPbsTerraShadows::getNumExtraPassTextures( const HlmsPropertyVec &properties,
-                                                         bool casterPass ) const
+    uint16 HlmsPbsTerraShadows::getNumExtraPassTextures(const HlmsPropertyVec& properties,
+                                                        bool casterPass) const
     {
-        return ( !casterPass && Hlms::getProperty( properties, PbsTerraProperty::TerraEnabled ) ) ? 1u
-                                                                                                  : 0u;
+        return (!casterPass && Hlms::getProperty(properties, PbsTerraProperty::TerraEnabled)) ? 1u
+            : 0u;
     }
     //-----------------------------------------------------------------------------------
     void HlmsPbsTerraShadows::propertiesMergedPreGenerationStep(
-        Hlms *hlms, const HlmsCache &passCache, const HlmsPropertyVec &renderableCacheProperties,
-        const PiecesMap renderableCachePieces[NumShaderTypes], const HlmsPropertyVec &properties,
-                                                       const QueuedRenderable &queuedRenderable )
+        Hlms* hlms, const HlmsCache& passCache, const HlmsPropertyVec& renderableCacheProperties,
+        const PiecesMap renderableCachePieces[NumShaderTypes], const HlmsPropertyVec& properties,
+        const QueuedRenderable& queuedRenderable)
     {
-         if( hlms->_getProperty( HlmsBaseProp::ShadowCaster ) == 0 &&
-            hlms->_getProperty( PbsTerraProperty::TerraEnabled ) != 0 )
+        if (hlms->_getProperty(HlmsBaseProp::ShadowCaster) == 0 &&
+            hlms->_getProperty(PbsTerraProperty::TerraEnabled) != 0)
         {
-            int32 texUnit = hlms->_getProperty( PbsProperty::Set0TextureSlotEnd );
-            hlms->_setTextureReg( VertexShader, "terrainShadows", texUnit - 1 );
+            int32 texUnit = hlms->_getProperty(PbsProperty::Set0TextureSlotEnd);
+            hlms->_setTextureReg(VertexShader, "terrainShadows", texUnit - 1);
         }
     }
     //-----------------------------------------------------------------------------------
-    void HlmsPbsTerraShadows::preparePassHash( const CompositorShadowNode *shadowNode, bool casterPass,
-                                               bool dualParaboloid, SceneManager *sceneManager,
-                                               Hlms *hlms )
+    void HlmsPbsTerraShadows::preparePassHash(const CompositorShadowNode* shadowNode, bool casterPass,
+                                              bool dualParaboloid, SceneManager* sceneManager,
+                                              Hlms* hlms)
     {
-        if( !casterPass )
+        if (!casterPass)
         {
 #if OGRE_DEBUG_MODE
             mSceneManager = sceneManager;
 #endif
 
-            if( mTerra && hlms->_getProperty( HlmsBaseProp::LightsDirNonCaster ) > 0 )
+            if (mTerra && hlms->_getProperty(HlmsBaseProp::LightsDirNonCaster) > 0)
             {
                 //First directional light always cast shadows thanks to our terrain shadows.
-                int32 shadowCasterDirectional = hlms->_getProperty( HlmsBaseProp::LightsDirectional );
-                shadowCasterDirectional = std::max( shadowCasterDirectional, 1 );
-                hlms->_setProperty( HlmsBaseProp::LightsDirectional, shadowCasterDirectional );
+                int32 shadowCasterDirectional = hlms->_getProperty(HlmsBaseProp::LightsDirectional);
+                shadowCasterDirectional = std::max(shadowCasterDirectional, 1);
+                hlms->_setProperty(HlmsBaseProp::LightsDirectional, shadowCasterDirectional);
             }
 
-            hlms->_setProperty( PbsTerraProperty::TerraEnabled, mTerra != 0 );
-            hlms->_setProperty( TerraProperty::ZUp, mTerra && mTerra->isZUp() );
+            hlms->_setProperty(PbsTerraProperty::TerraEnabled, mTerra != 0);
+            hlms->_setProperty(TerraProperty::ZUp, mTerra && mTerra->isZUp());
         }
     }
     //-----------------------------------------------------------------------------------
-    uint32 HlmsPbsTerraShadows::getPassBufferSize( const CompositorShadowNode *shadowNode,
-                                                   bool casterPass, bool dualParaboloid,
-                                                   SceneManager *sceneManager ) const
+    uint32 HlmsPbsTerraShadows::getPassBufferSize(const CompositorShadowNode* shadowNode,
+                                                  bool casterPass, bool dualParaboloid,
+                                                  SceneManager* sceneManager) const
     {
         return (!casterPass && mTerra) ? 32u : 0u;
     }
     //-----------------------------------------------------------------------------------
     float* HlmsPbsTerraShadows::prepareConcretePassBuffer(const Ogre::CompositorShadowNode* shadowNode, bool casterPass, bool dualParaboloid, Ogre::SceneManager* sceneManager, float*& passBufferPtr)
     {
-        if( !casterPass && mTerra )
+        if (!casterPass && mTerra)
         {
             const float invHeight = 1.0f / mTerra->getHeight();
-            const Vector3 &terrainOrigin = mTerra->getTerrainOriginRaw();
-            const Vector2 &terrainXZInvDim = mTerra->getXZInvDimensions();
+            const Vector3& terrainOrigin = mTerra->getTerrainOriginRaw();
+            const Vector2& terrainXZInvDim = mTerra->getXZInvDimensions();
             *passBufferPtr++ = -terrainOrigin.x * terrainXZInvDim.x;
             *passBufferPtr++ = -terrainOrigin.y * invHeight;
             *passBufferPtr++ = -terrainOrigin.z * terrainXZInvDim.y;
@@ -141,17 +141,17 @@ namespace Ogre
         return passBufferPtr;
     }
     //-----------------------------------------------------------------------------------
-    void HlmsPbsTerraShadows::hlmsTypeChanged( bool casterPass, CommandBuffer *commandBuffer,
-                                               const HlmsDatablock *datablock, size_t texUnit )
+    void HlmsPbsTerraShadows::hlmsTypeChanged(bool casterPass, CommandBuffer* commandBuffer,
+                                              const HlmsDatablock* datablock, size_t texUnit)
     {
-        if( !casterPass && mTerra )
+        if (!casterPass && mTerra)
         {
-            Ogre::TextureGpu *terraShadowTex = mTerra->_getShadowMapTex();
+            Ogre::TextureGpu* terraShadowTex = mTerra->_getShadowMapTex();
 
             //Bind the shadows' texture. Tex. slot must match with
             //the one in HlmsPbsTerraShadows::propertiesMergedPreGenerationStep
-            *commandBuffer->addCommand<CbTexture>() = CbTexture( texUnit++, terraShadowTex,
-                                                                 mTerraSamplerblock );
+            *commandBuffer->addCommand<CbTexture>() =
+                CbTexture((uint16)texUnit++, terraShadowTex, mTerraSamplerblock);
         }
     }
-}
+} // namespace Ogre
